provide a valid listing for compound cube
The previous listings defined for CompoundCubes/Doctors, CompoundCubes/Patients, and CompoundCubes/CityRainfall did not have identical column specs, as required in order for listings to be run on the CompoundCubes/CompoundCube subject area. The new CompoundListing in each of the three cubes should correct this, though I have encountered errors (which may be an IRIS bug) when testing it on CompoundCubes/CompoundCube if there is a dimension on rows or columns.